+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* the --output flag in `lsystem` now has a default value of `out.svg`.
+
 ### Removed
 
 ## [2.0.0]